requirements.txt
future==0.15.2
oauthlib==2.0.0
pkg-resources==0.0.0
python-twitter==3.1
requests==2.11.1
requests-oauthlib==0.7.0
